requirements.txt
gitpython==3.1.41
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.4